Make buildable false by default
Most packages put in packages.yaml will be used to link external/system
packages which the user will not likely want to build again. Having to
remember to add buildable: False to each entry can be a pain. This
commit, makes buildable: False the default and if the user wants to
build an external package they can explicitly add buildable: True.