Turn build_env_compilers into context manager
This ideas has been discussed before and rather than having a decorator
that wraps around a function to be executed in a frontend environment it
is better to have it as a context manager. The idea here is that
arbitrary code can be executed in a child process that sets up the
frontend. Each platform should know how to handle their frontend env.