Add wrapper that executes func in subprocess
This decorator function wraps an executable (like ./b2 in boost and
bootstrap in cmake) and executes that function in subprocess where the
frontend environment is setup. The reason for this is that some
executables are built using the backend compilers, but since we are on a
different architecture these executables fail. As a workaround, we
execute the build for thr frontend, so that the resulting executable can
run on a login node, using our decorator. It looks deeply nested because
we need to first make sure it is in module scope and then we make sure
to wrap our executable and then when we run it, we can run it using
multiprocessing.Process.