Remove conflict and throw error building on Cray
Cray compiler wrappers were likely causing unzip to not build on Cray
machines, however, any package that depends on unzip will then be
unbuildable (eg vim). By removing conflict we can at least throw an
error if the user tries to build it for the backend but also allows the
user to install it if they are just using the regular compilers.